Switch to 32 bit ints, add perf analysis for it

+++ src/pymatching/fill_match/ints.h
@@ -12,20 +12,20 @@ namespace pm {
 typedef uint64_t obs_int;
 
 /// This type is used to store the weight of an edge.
-typedef uint16_t weight_int;
+typedef uint32_t weight_int;
 
 /// This type is used to store the potentially-negative weight of an edge.
 /// It is used when loading graphs in order to support negative edge weights.
 /// However, negative edge weights are handled in pre- and post-processing, rather
 /// than in the blossom algorithm itself, so the `weight_int' type is used instead for data
 /// structures used within the blossom algorithm.
-typedef int16_t signed_weight_int;
+typedef int32_t signed_weight_int;
 
 /// This type is used to represent absolute times, accumulated times, and accumulated distances.
 /// It is important that it be signed because, for example, it's possible to compute potential
 /// collision times that are in the past while considering whether a collision will occur in the
 /// future or not.
-typedef int32_t cumulative_time_int;
+typedef int64_t cumulative_time_int;
 
 /// This type is used to represent the total weight of the MWPM solution. It is important that it
 /// is 64-bit since in general we expect the total solution weight to grow linearly with the number of
@@ -36,7 +36,7 @@ typedef int64_t total_weight_int;
 /// This type is used to represent timestamps that are near the current time. To widen a cyclic
 /// time to an absolute time, find the absolute time which is closest to the current time and
 /// equal to the cyclic time when truncated..
-typedef cyclic<uint16_t> cyclic_time_int;
+typedef cyclic<uint32_t> cyclic_time_int;
 
 }  // namespace pm
